Bail Polo client on bad data or channel ID
- Noticing that the socket seems to occasionally miss some data, not
sure if this is primarily on my end of polo's but nothing is immediately
indication to me that the service is failing to parse these message - my
guess is that they were either dropped in transit or polo's data feed is
generally flaky. For now going to open a fresh connection when service
receives unknown/bad data or heartbeats (they only get sent after longer
periods with no messages, I'm thinking these may also be indicative of a
faulty data feed. On reconnect the book is re-generated so this should
help resolve issues with hanging orders.